Move curated extensions to external extensions util
